Avoid implicit narrowing conversions

+++ qupath-core-processing/src/main/java/qupath/lib/analysis/features/LocalBinaryPatterns.java
@@ -233,7 +233,7 @@ public class LocalBinaryPatterns {
         double w4 =      tx  *      ty;
         // Compute the value
         // By checking weights first, this avoids requesting out-of-image pixels (assuming x & y were ok)
-        float value = 0;
+        double value = 0;
         if (w1 > 0)
         	value += w1*img.getValue(fx, fy);
         if (w2 > 0)
@@ -242,7 +242,7 @@ public class LocalBinaryPatterns {
         	value += w3*img.getValue(fx, cy);
         if (w4 > 0)
         	value += w4*img.getValue(cx, cy);
-        return value;
+        return (float)value;
 	}
